Refactor navbarbuttons to be navbar
